requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
google-generativeai>=0.5
python-dotenv>=1.0
pillow>=10.0
whitenoise>=6.6
